piper-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
piper-tts==1.2.0